Add 'Badges' functionality and do routing corrections
This commits introduces a new 'Badges' page to the Ionic application in the 'src\app\badges' directory. The Badges page HTML,  SCSS, specs and typescript files were added. This page contains different badges with icons and text to demonstrate the use of Ionic badges. It also modifies 'tabs' to include a routing button for the 'badges' page.

To integrate the 'Badges' page into the app, 'tabs.routing.ts' was edited to bring 'badges' page into tabs navigation. 'personCircle' icon was also added in 'tabs.page.ts' and used it for the login button in 'tabs.page.html'.

A change was made to the navigation in 'informe-dados.page.ts' to correctly route back to "/tabs/home" instead of "/home".

Lastly, the README was updated with the project's detailed documentation. Importantly, it explained the new 'Badges' page and updated the project structure.

These changes were made to enhance the app's functionality and also correct routing in the app.